feat: PageIndex 컴포넌트 로직을 커스텀 훅으로 분리 및 수정
- page의 초기값으로 1이 아닌 initialPage 를 입력:
searchParams의 변경으로 리렌더링이 일어날 때 초기값인 1로 초기화되어 페이지가 1로 변경되는 문제점 발견
searchParams의 skip값을 가져와서 skip값에 따라 초기 페이지값을 결정
